coin_trade/pyproject.toml
[tool.poetry]
name = "coin-trade"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
requests = "2.28.1"
pandas = "1.5.0"
numpy = "1.26.3"
python-telegram-bot = "13.14"
pyjwt = "2.5.0"
pyinstaller = "5.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
